Feat : 서버 액션을 관리하는 actions.ts 파일 추가
